// src/types/user.ts
export type UserType = {
    id: string;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    address: string;
    jobTitle: string;
    role: string;
    status: string;
    bio: string;
    imageUrl: string;
    lastLogin: string;
    createdAt: string;
    updatedAt: string;
};